+++ Sodaq_Universal_Tracker/Network.h
@@ -54,7 +54,7 @@ public:
 
     bool init(Uart& modemStream, DataReceiveCallback callback, uint32_t(*getNow)(), InitConsoleMessages messages, InitJoin join);
 
-    uint8_t transmit(uint8_t* buffer, uint8_t size, uint32_t rxTimeout);
+    size_t transmit(uint8_t* buffer, size_t size, uint32_t rxTimeout);
 
     void loopHandler();
